pyproject.toml
[tool.poetry]
name = "ai-matchmaker"
version = "0.1.0"
description = ""
authors = ["Dev Aggarwal <devxpy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dash = "^2.9.3"
openai = "^0.27.7"
pandas = "^2.0.1"
plotly = "^5.14.1"
umap-learn = "^0.5.3"
furl = "^2.1.3"
google-api-python-client = "^2.87.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
sklearn = "^0.0.post5"
gunicorn = "^20.1.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
